Add support for thinjails
When installing an AppJail, the user can install it as a thinjail
instead of the default, thickjail. The combination of tinyjail
and thinjails results in a very stripped down jail.Add support for thinjails

When installing an AppJail, the user can install it as a thinjail
instead of the default, thickjail. The combination of tinyjail