Fix environment variables
When environment variables containing harmful characters, such as
quotes, are passed, this is not correctly escaped when `su(1)` logs into
another user because the shell may be other than `sh(1)` (the only shell
required by RUN to work properly).

This was tested using the `root` user in a FreeBSD jail. This user has
`csh(1)` as the shell, so the escape is very different than `sh(1)`.