Add --maintain-env parameter
Some applications work fine even if the environment does not match the
current user, so --maintain-env is used to use -m in `su(1)`. This is
very useful since `su(1)` does not work if the user does not have an
existing home directory and a shell when emulating a full login (by
default using -l).