Install with the correct permissions
When using an umask of 0022 all files are installed correctly on the
system, however when using an umask of 0027 (my case) scripts like
bin/appjail-config or bin/appjail cannot read
share/appjail/files/config.conf so an error is generated when loading
it.

This happens when cloning the repository when an umask 0027 or other
than 0022. To fix this, it is worth installing all files and directories
with specific permissions.