Fix incorrect detection of mounted points
A problem can arise when one jail has a name like `hello` and another
jail with a name like `hello-builder`. The problem is when a mounted
point is detected in the `hello` jail but
`share/appjail/lib/mount:lib_mountpoint_mounted` will show mount points
in `hello-builder` which is incorrect. A substring comparison is used to
fix the problem.