Fix invalid update check for ajspec file
Interrupting with e.g. CTRL-C when getting an image may cause an invalid
check of ajspec updates after retrying the same operation, because the
ajspec obtained when importing is copied before getting the image. To
solve this, the ajspec file obtained is copied after getting the image,
almost at the end.