Improve killing processes in healthcheck & startup
Simply killing a process is not technically correct because the process
after a while may be another process created by another process and not
by AppJail, so errors may occur. To fix this, only processes created by
AppJail will be killed, this is done by comparing the parent pid with
the parent pid of the process we want to kill.