Fix <defunct> process in appjail-dns
When sending a SIGTERM signal to appjail-dns, there will be one process
left, that process is the `sleep(1)` command, which results in a minute
or a little more waiting for this process to finish. To fix this, when
sending a SIGTERM signal to appjail-dns, a SIGTERM signal is sent to the
`sleep(1)` command.