Add fetch src method, Add deleteOld command, Fix update jail parameters
fetch src:

This implementation tries to be as flexible as possible for users who
want to take advantage of the source tree, but not as flexible as the
`fetch empty` method.

Both `thick` and `thin` jail types are allowed, but as a note for anyone
interested in using thickjails is that AppJail checks if the dummy file
`.from_src` exists in the release directory of that jail. This creates a
link between the release and the jail, which is relevant when exporting
the jail to another system. This is relevant when the user wants to
update or export the jail. For updates, the compilation will be affected
depending on the result and the options used by `fetch src`. For exports,
the target system will not know if the jail comes from a source tree or not,
unless the release directory of that system has the dummy file (probably
through an export/import from the release directory).

update jail:

* shift is not executed after getopts finishes its execution, so the
  jail name is incorrect.
* case...esac do not have a conditional to check if a parameter is
  correct or not.