Fix current directory when executing initscripts
When executing an initscript, the current directory should be the last
current directory when processing the main Makejail, but it is the
current directory where the initscript is located, which can break
stages using some instructions like `COPY` or `CMD --local`. The build
stage is not affected since it does not use initscripts.