Fix functions and modules that may hang
These include:

* fetch
* login
* lib_check_signal
* lib_check_rctl*
* lib_check_path_traversal*

These functions naively pass an argument to `grep(1)` relying on the
`-F` flag to compare each entry. The problem is that if the user passes
an invalid entry as `-a`, problems such as hangings or the like may occur.
To fix this, simply add `--` before the pattern to told to `getopt(3)`
to stop processing further options.