Ignore SIGINT for initscripts and the buildscript
The child processes are terminated successfully, but in an interactive
session (using AppJail on a console) a CTRL-C (so a SIGINT) can close
child processes like the buildscript or an initscript (both are really a
`sh(1)` script). To work around this, SIGINT is ignored, but all other
signals are not, so kill_tree.sh of the parent process will terminate
the remaining processes afterwards.