Don't configure port forwarding unless the user explicitly requests it
When creating an OCI container and the image has ports configured
to be forwarded, AppJail may generate an error if the user uses
an option other than `virtualnet`, since only with that option `expose`
takes effect. For example, if the user uses a combination of `alias`
and `ip4_inherit`, AppJail will generate an error if the image wants
to expose port `5055`. Now the user must set `container=expose` for
it to take effect.