Set failure policy to "continue" in dyndns hook
Since the purpose of this hook is to update the IPv4 address
immediately upon deploying a service, it works as expected
in these situations. In the case of startup, it may or may
not work for various reasons, but mainly due to a DNS issue,
because dyndnsd has not yet started, or because NGINX has
not yet started. This applies in cases where all services
are down at the same time (such as a data center outage)
and start at the same time. This should not affect cases
where dyndnsd and NGINX (if dyndnsd is behind a reverse proxy)
have good uptime, but setting the fail policy to "continue"
is a good choice for this worst-case scenario, as the IPv4
address should not change frequently or even change at all.